Change FireworkData to be two separate classes, to more easily represent firework charges
